Occlum musl libc is compatible with native OS
Before this commit, programs that are compiled with Occlum toolchain and thus
linked with Occlum musl libc cannot run on the native OS (e.g., Linux). This
makes building (cross-compiling) and testing (cannot see if it works on Linux)
difficult. With new commit, we enable Occlum musl libc to detect which platform
that the program is running. If it is on Occlum, call Occlum sysclal;
otherwise, call native syscall.